Controller Nominally Finished
The combined Nufern and Orbits labview controller nominally works. There
are still issues with a few things including the device power-down
procedure (both have to be hardware disabled to fully turn off, or we're
not sending the right controls to turn off), as well temperature
management of the Orbits. It appears that a user set power level fights
against automatic temperature stabilization and when the controller is
adjusting temperature, all the Orbits readouts go crazy. Still works
though...